Implementacion de cookies, menos en comentarios

+++ frontend/pages/asamblea/ver/[asamblea].jsx
@@ -15,18 +15,33 @@ import {
     ModalCloseButton,
     useDisclosure
   } from '@chakra-ui/react'
-
-
+import {checkToken} from '../../../data/usuario'
+const jwt = require("jwt-simple")
 
 export async function getServerSideProps(context){
     try {
-        const response = await axios.get(`${process.env.API_URL}/asamblea/search/${context.params.asamblea}`)
-        return{
-            props:{
-                asambleaId: response.data
+        const res = await checkToken(context.req.headers.cookie)
+        const decode = jwt.decode(context.req.cookies.token,process.env.SECRET_KEY)
+            if (res.status === 200){
+                const response = await axios.get(`${process.env.API_URL}/asamblea/search/${context.params.asamblea}`)
+                return{
+                    props:{
+                        asambleaId: response.data,
+                        existe: res.config.headers.cookie,
+                        rol:decode.rol
+                    }
+                }
+            }else{
+                console.log("No hay token")
+                return{
+                    redirect: {
+                        destination: "/",
+                        permanent: false
+                    }
+                }
             }
-        }
     } catch (error) {
+        console.log("ERROR",error)
         return{
             redirect:{
                 destination: '/asamblea/ver',
@@ -39,7 +54,7 @@ export async function getServerSideProps(context){
 const asamblea = (data) => {
     const router = useRouter()
     const [asambleas] = useState(data)
-    const[values, setValues] = useState({})
+    const[values, setValues] = useState({rolUsuario: 'admin'})
     const { isOpen, onOpen, onClose } = useDisclosure()
 // ------------------------------------------------------------------------------------------------------------------------------------------------------------
     //moi
@@ -80,31 +95,41 @@ const asamblea = (data) => {
 
 
     const eliminarAsamblea = async () =>{
-        try {
-            const response = await axios.delete(`${process.env.API_URL}/asamblea/delete/${asambleas.asambleaId._id}`,{data: values })
-            if (response.status === 200){
-                Swal.fire({
-                    title: 'Asamblea eliminada',
-                    text: 'La asamblea se ha eliminado con exito',
-                    icon: 'success',
-                    confirmButtonText: 'Ok'
-                }).then((result)=>{
-                    if (result.isConfirmed){
-                        router.push("/asamblea/ver")
-                    }
-                })
-            }
-            }catch (error) {
-                Swal.fire({
-                    title: 'Error',
-                    text: `La asamblea no se ha podido eliminar`,
-                    icon: 'error',
-                    confirmButtonText: 'Ok'
-                })
-            }
+        if (data.rol==='admin'){
+            try {
+                const response = await axios.delete(`${process.env.API_URL}/asamblea/delete/${asambleas.asambleaId._id}`,{data: values })
+                if (response.status === 200){
+                    Swal.fire({
+                        title: 'Asamblea eliminada',
+                        text: 'La asamblea se ha eliminado con exito',
+                        icon: 'success',
+                        confirmButtonText: 'Ok'
+                    }).then((result)=>{
+                        if (result.isConfirmed){
+                            router.push("/asamblea/ver")
+                        }
+                    })
+                }
+                }catch (error) {
+                    Swal.fire({
+                        title: 'Error',
+                        text: `La asamblea no se ha podido eliminar`,
+                        icon: 'error',
+                        confirmButtonText: 'Ok'
+                    })
+                }
+        }else{
+            Swal.fire({
+                title: 'Error',
+                text: `No tiene los permisos para eliminar asambleas`,
+                icon: 'error',
+                confirmButtonText: 'Ok'
+            })
+        }
     }
 
     const desvincularArchivos = async () =>{
+        if (data.rol==='admin'){
         try {
             const response = await axios.delete(`${process.env.API_URL}/file/deleteAll/${asambleas.asambleaId._id}`)
             if (response.status === 200){
@@ -125,6 +150,14 @@ const asamblea = (data) => {
                 confirmButtonText: 'Ok'
             })
         }
+        }else{
+            Swal.fire({
+                title: 'Error',
+                text: `No tiene los permisos para eliminar archivos asociados`,
+                icon: 'error',
+                confirmButtonText: 'Ok'
+            })
+        }
     }
 
 // ------------------------------------------------------------------------------------------------------------------------------------------------------------
@@ -140,25 +173,34 @@ const asamblea = (data) => {
     }
 
     const eliminarArchivos = async (id_archivo, id_asamblea) =>{
-        try {
-            //AGREGAR MODALs
-            const response = await axios.delete(`${process.env.API_URL}/file/delete/${id_archivo}/${id_asamblea}`)
-            if (response.status === 200){
+        if (data.rol==='admin'){
+            try {
+                //AGREGAR MODALs
+                const response = await axios.delete(`${process.env.API_URL}/file/delete/${id_archivo}/${id_asamblea}`)
+                if (response.status === 200){
+                    Swal.fire({
+                        title: 'Archivo eliminado',
+                        text: 'El archivo se ha eliminado con exito',
+                        icon: 'success',
+                        confirmButtonText: 'Ok'
+                    }).then((result)=>{
+                        if (result.isConfirmed){
+                            router.reload()
+                        }
+                    })
+                }
+                }catch (error) {
                 Swal.fire({
-                    title: 'Archivo eliminado',
-                    text: 'El archivo se ha eliminado con exito',
-                    icon: 'success',
+                    title: 'Error',
+                    text: `El archivo no se ha podido eliminar`,
+                    icon: 'error',
                     confirmButtonText: 'Ok'
-                }).then((result)=>{
-                    if (result.isConfirmed){
-                        router.reload()
-                    }
                 })
             }
-            }catch (error) {
+        }else{
             Swal.fire({
                 title: 'Error',
-                text: `El archivo no se ha podido eliminar`,
+                text: `No tiene los permisos para eliminar archivos`,
                 icon: 'error',
                 confirmButtonText: 'Ok'
             })
@@ -184,6 +226,7 @@ const asamblea = (data) => {
 
     const subirArchivos = async (id_asamblea) => {
 
+        if (data.rol==='admin'){
         const { value: file } = await Swal.fire({
         title: 'Seleccionar Archivo',
         input: 'file',
@@ -231,6 +274,14 @@ const asamblea = (data) => {
                 confirmButtonText: 'Ok'
             })
         }
+        }else{
+            Swal.fire({
+                title: 'Error',
+                text: `No tiene los permisos para subir archivos`,
+                icon: 'error',
+                confirmButtonText: 'Ok'
+            })
+        }
     }
 
 
@@ -238,7 +289,7 @@ const asamblea = (data) => {
 
     return (
         <Box>
-            <Arriba/>
+            <Arriba token={data.existe}/>
             <Container maxW="container.xl" >
                 <Heading my={15} textAlign={"center"}> {asambleas.asambleaId.name}</Heading>
 
@@ -248,7 +299,7 @@ const asamblea = (data) => {
                         <Button w={"full"} colorScheme={"green"} onClick={() => router.push(`/asamblea/editar/${asambleas.asambleaId._id}`)}>Editar</Button>
                         <Button w={"full"} colorScheme={"teal"} onClick={() => router.push(`/asistencia/ver/${asambleas.asambleaId._id}`)}>Ver asistencias asamblea</Button>
                         <Button w={"full"} colorScheme={"red"} onClick={onOpen}>Eliminar asamblea</Button>
-                        
+
                         <Modal isOpen={isOpen} onClose={onClose}>
                             <ModalOverlay/>
                                 <ModalContent>
@@ -263,12 +314,12 @@ const asamblea = (data) => {
                         </Modal>
 
                     </HStack>
-                <RadioGroup >
+                {/* <RadioGroup >
                     <HStack spacing='24px' >
                     <Radio value='user' onChange={onChange} name={"rolUsuario"}>user</Radio>
                     <Radio value='admin' onChange={onChange} name={"rolUsuario"}>admin</Radio>
                     </HStack>
-                </RadioGroup>
+                </RadioGroup> */}
                 <Divider/>
                 <Stack w={"full"} py={10}>
                     <ShowInfo tag="Nombre" data={asambleas.asambleaId.name} />
